HEC-470: Extract CommandLoader from CRUD capability to stay under 200 lines
